don't need a separate interface for convenience methods for handlers
