+++ orca-queue/src/main/kotlin/com/netflix/spinnaker/orca/q/handler/ConfigurationErrorHandler.kt
@@ -17,6 +17,7 @@
 package com.netflix.spinnaker.orca.q.handler
 
 import com.netflix.spinnaker.orca.ExecutionStatus.TERMINAL
+import com.netflix.spinnaker.orca.pipeline.persistence.ExecutionRepository
 import com.netflix.spinnaker.orca.q.CompleteExecution
 import com.netflix.spinnaker.orca.q.ConfigurationError
 import com.netflix.spinnaker.orca.q.MessageHandler
@@ -27,7 +28,8 @@ import org.springframework.stereotype.Component
 @Component
 open class ConfigurationErrorHandler
 @Autowired constructor(
-  override val queue: Queue
+  override val queue: Queue,
+  override val repository: ExecutionRepository
 ) : MessageHandler<ConfigurationError> {
 
   override val messageType = ConfigurationError::class.java
