+++ orca-queue/src/test/kotlin/com/netflix/spinnaker/orca/q/handler/ConfigurationErrorHandlerSpec.kt
@@ -18,6 +18,7 @@ package com.netflix.spinnaker.orca.q.handler
 
 import com.netflix.spinnaker.orca.ExecutionStatus.TERMINAL
 import com.netflix.spinnaker.orca.pipeline.model.Pipeline
+import com.netflix.spinnaker.orca.pipeline.persistence.ExecutionRepository
 import com.netflix.spinnaker.orca.q.*
 import com.nhaarman.mockito_kotlin.mock
 import com.nhaarman.mockito_kotlin.reset
@@ -29,8 +30,9 @@ import org.jetbrains.spek.api.dsl.it
 class ConfigurationErrorHandlerSpec : Spek({
 
   val queue: Queue = mock()
+  val repository: ExecutionRepository = mock()
 
-  val handler = ConfigurationErrorHandler(queue)
+  val handler = ConfigurationErrorHandler(queue, repository)
 
   fun resetMocks() = reset(queue)
